Updated README and added more fn features

+++ index.js
@@ -11,7 +11,7 @@ const token = '<your token here>'; //if running locally
 
 //API configs
 const fortniteConfig = {
-    apikey: "fortnite-api.com API Key here>",
+    apikey: "<fortnite-api.com API Key here>",
     language: "en",
     debug: true
 };
@@ -42,6 +42,9 @@ bot.on('message', msg => {
             if (!msg.guild) return msg.reply('Please use this bot in a guild.')
             if (!msg.member.hasPermission('ADMINISTRATOR', true)) return msg.channel.send('THIS IS A MOD-ONLY COMMAND, YOU DO NOT HAVE PERMISSIONS TO USE THIS COMMAND. THIS ACTION WILL BE LOGGED').then(msg.guild.channels.cache.get(logChannel).send(`${msg.author.tag} used the mod-only command (clear) in #${msg.channel.name}`))
             if (!args[1]) return msg.reply('Please specify a number of messages that you would like to delete')
+            if (isNaN(args[1])) return msg.reply('Please give a number.')
+            if (args[1] > 100) return msg.reply('You cannot delete more than 100 messages at a time.')
+            if (args[1] < 1) return msg.reply('You must delete at least one message.')
             msg.delete({ timeout: 1000, reason: 'Hides command so normal users cannot see.' })
             msg.channel.bulkDelete(args[1])
             msg.guild.channels.cache.get(logChannel).send(`@${msg.author.tag} deleted ${args[1]} messages in #${msg.channel.name}`)
@@ -81,7 +84,8 @@ bot.on('message', msg => {
             if (!msg.guild) return msg.reply('Please use this bot in a guild.')
             let helpEmbed = new Discord.MessageEmbed()
                 .setTitle('PickleRick Bot Help')
-
+                .addField('Information', 'Find out information about members using pr!minfo @..., or do pr!cinfo or pr!sinfo to get channel and server information respectively.')
+                .addField('Having Fun', 'To have fun, get a random image from the internet on the default subject of memes with pr!gi <subject (optional and has to have NO spaces)>. Or find out stuff about Fortnite, get started by doing pr!fn help')
                 .setThumbnail(msg.author.displayAvatarURL());
             msg.author.send(helpEmbed)
             msg.delete({ timeout: 1000, reason: 'Hides command so normal users cannot see.' })
@@ -271,7 +275,7 @@ bot.on('message', msg => {
             }
             let mcHelpEmbed = new Discord.MessageEmbed()
                 .setTitle('Minecraft Online Server Status Command Help')
-                .addField('Command Format', '$mcserverstatus <server IP> ')
+                .addField('Command Format', 'pr!mc <server IP> ')
                 .setThumbnail('https://thumbs.dreamstime.com/b/minecraft-logo-online-game-dirt-block-illustrations-concept-design-isolated-186775550.jpg');
             if (!serverIP) return msg.channel.send(mcHelpEmbed)
             msg.reply('Please wait a second while I get the server\'s status')
@@ -320,7 +324,8 @@ bot.on('message', msg => {
                     .setTitle('Fortnite Commands Help')
                     .addField('Fortnite Map', 'You can get the fortnite map using the command pr!fn map')
                     .addField('Fortnite User Stats', 'Get stats on a specifc user from online via pr!fn stats <username, e.g Lachy>. Additionally, you can add an account type at the back, like pr!fn stats PWR%Lachy epic, to make account-specific queries. Accepted account types are epic, psn and xbl. (case-sensitive).')
-                    .addField('Fortnite Shop', 'Get information about the items in the Item Shop by pr!fn shop');
+                    .addField('Fortnite Shop', 'Get information about the items in the Item Shop by pr!fn shop')
+                    .setThumbnail('https://thumbs.dreamstime.com/b/fortnite-white-vector-logo-black-textured-background-online-game-editorial-illustration-144436055.jpg');
                 msg.channel.send(fnHelpEmbed)
             }
 
@@ -341,11 +346,9 @@ bot.on('message', msg => {
                 if (!username) return msg.reply('Please give the user\'s name')
                 let accountType = args[3]
                 if (!accountType) {
-                    console.log(username)
                     let params = { name: username, image: 'all' }
                     getFnStats(msg, params)
                 } else if (accountType === 'epic' || accountType === 'psn' || accountType === 'xbl') {
-                    console.log(username)
                     let params = { name: username, accountType: accountType, image: 'all' }
                     getFnStats(msg, params)
                 } else {
@@ -354,12 +357,61 @@ bot.on('message', msg => {
                 }
             } else if (fnParam === 'troubleshoot') {
                 let fnTrblshtEmbed = new Discord.MessageEmbed()
-                .setTitle('Fortnite Command Error Troubleshooting')
-                .addField('pr!fn map', 'If you are having trouble in getting the map, it is probably because Fortnite is mid-update or the API servers are down. Please try again in a few hours.')
-                .addField('pr!fn stats <user> <accountType>', 'If you are getting this error, it is probably because the user you want to get stats on doesn\'t exist in the query. Also, ensure that you replace any spaces in the name with %, as this also can cause errors. The reason could also be that you typed an invalid account type, accepted account types are epic, psn and xbl (case-sensitive). If not, try, if you had made a account-based query, not making one or if you did make an account-based query, try making a global one.')
-
+                    .setTitle('Fortnite Command Error Troubleshooting')
+                    .addField('pr!fn map', 'If you are having trouble in getting the map, it is probably because Fortnite is mid-update or the API servers are down. Please try again in a few hours.')
+                    .addField('pr!fn stats <user> <accountType>', 'If you are getting this error, it is probably because the user you want to get stats on doesn\'t exist in the query. Also, ensure that you replace any spaces in the name with %, as this also can cause errors. The reason could also be that you typed an invalid account type, accepted account types are epic, psn and xbl (case-sensitive). If not, try, if you had made a account-based query, not making one or if you did make an account-based query, try making a global one.')
+                    .addField('pr!fn cc <creator code>', 'Ensure you are typing the creator code correctly as the code is case-sensitive. Also, if the code has spaces, ensure tha you replace them with %, like: Royal%Warrior')
                 msg.channel.send(fnTrblshtEmbed)
-            } 
+            } else if (fnParam === 'cc') {
+                let name = args[2]
+                if (!name) return msg.reply('Please provide the creator code to get details on.')
+                name = name.replace('%', ' ')
+                if (typeof name !== 'string') return msg.reply('Please provide a proper creater code')
+                msg.reply('Please wait a moment while I get the details on the creator code.')
+                fortniteStats.CreatorCode(name)
+                    .then(res => {
+                        let fnCCEmbed = new Discord.MessageEmbed()
+                            .setTitle(`${name} Creator Code Details`)
+                            .addField('Code', `${res.data.code}`)
+                            .addField('Owner Account ID and Name', `ID: ${res.data.account.id}, Name: ${res.data.account.name}`)
+                            .addField('Is Verified', `${res.data.verfiied}`);
+                        msg.channel.send(fnCCEmbed)
+                    }).catch(err => {
+                        msg.channel.send('An error occurred in getting the creator code details. Type pr!fn troubleshoot for help.')
+                        console.log(err)
+                    })
+            } else if (fnParam === 'news') {
+                msg.reply('Please wait a moment while I get the latest Battle Royale news.')
+                fortniteStats.NewsBR('en')
+                .then(res => {
+                    msg.channel.send(res.data.image)
+                }).catch(err => {
+                    msg.channel.send('An error occurred in getting the news. The API servers could be down. Try again in a few minutes.')
+                    console.log(err)
+                })
+            } else if (fnParam === 'cos') {
+                let name = args[2]
+                if (!name) return msg.reply('Please give the name of the cosmetic you want information on.')
+                if (typeof name !== 'string') return msg.reply('Please give the proper name of the cosmetic.')
+                name = name.replace('%', ' ')
+                msg.reply('Please wait a moment while I search for the Fortnite cosmetic with the given name.')
+                fortniteStats.CosmeticsSearch({name: name})
+                .then(res => {
+                    let fnCosEmbed = new Discord.MessageEmbed()
+                    .setTitle(`${name} Cosmetic Details`)
+                    .addField('Name And ID', `Name: ${name}, ID: ${res.data.id}`)
+                    .addField('Description', res.data.description)
+                    .addField('Rarity', res.data.rarity.value)
+                    .addField('Introduction', `Chapter ${res.data.introduction.chapter}, Season ${res.data.introduction.season}`)
+                    .addField('Added On', `${res.data.added}`);
+                    msg.channel.send(res.data.images.featured)
+                    msg.channel.send(fnCosEmbed)
+                }).catch(err => {
+                    msg.channel.send('An error occurred in getting the cosmetic\'s name. Ensure that you typed it correctly and that you replaced any spaces with %')
+                    console.log(err)
+                })
+            }
+            break;
     }
 })
